requirements.txt
requests==2.28.1
python-dotenv==0.19.2
boto3==1.26.0
authlib==1.2.1  # For OAuth2 if needed